requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==19.3.0
Babel==2.8.0
board==1.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.0.8
django-oauth-toolkit==1.3.2
djangorestframework==3.11.0
idna==2.10
more-itertools==8.4.0
oauthlib==3.1.0
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
py==1.9.0
pyparsing==2.4.7
pytest==5.4.3
pytz==2020.1
requests==2.24.0
reverse==0.1.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.2.5
